Remove the check for an already existing backend
In our codebase, we have the need to retry `isSupported( )`
calls. This is due to, sometimes, the extension taking a while to
respond to the U2F Api. However, since the `_backend` is a promise
object, after the first call, it stays as resolved for the whole
lifecycle of the library.

This commit removes the null check on the `getBackend( )` function, in
order to force the creation of a new Promise every time. This allows
the user to create a promise chaining mechanism to retry the
`isSupported( )` validation.